Fixing Memory Leak BaseSwap

+++ utilities/base_swap.c
@@ -44,7 +44,7 @@ const char* convert_2bits_to_32(const char* number_to_convert){
 
     int result_length = sizeof(numbers_base_10)/sizeof(short int);
 
-    char* result = (char*) calloc(result_length, sizeof(char));
+    char* result = (char*) calloc((size_t) result_length, sizeof(char));
     int j;
 
     if(result != NULL)
